Fix fish(6)'s dealing from the deck.
Before this, it would pick one of the ranks with cards remaining,
uniformly, and hand out one of that card.  This makes it more likely
than it should be that it'll deal ranks for which the deck is low on
cards (though, fortunately, it does stop dealing a rank once that rank
is exhausted).

With this change, it behaves like a real deck (well, assuming the RNG
works properly).