Kludge around a backgammon(7) bug that gets the number of possible moves wrong.
Here's an example.  I tried to move 19/1/4 here and got this rejection:

_____________________________________________________
|                       |   |                       |
|13  14  15  16  17  18 |   |19  20  21  22  23  24 |
|     w                 |   | r           r   w   r | r
|     w                 |   | r           r   w   r | r
|                       |   | r           r       r | r
|                       |   |                     r | r
|                       |   |                     r |
|                       |BAR|                       |
|                       |   |                       |
|                       |   |                       |
|                       |   |                     w |
|                       |   | w   w           w   w |
| w           w         |   | w   w           w   w |
|12  11  10   9   8   7 |   | 6   5   4   3   2   1 |
|_______________________|___|_______________________|
Player is red, computer is white.  Score:  Red 10, White 4

White rolls 2 4 and moves 14-12,13-9.
Red rolls 4 1.                  Move:
Red can only make 1 move.

This is obviously wrong.  On buiding a scaffolding program, I found
that if the dice showed 1 4 instead of 4 1, the code worked, thus
indicating there definitely is a bug.  This is not the right fix; I
suspect the right fix is to do something with dice swapping in the
"if (m == 2 && i)" block about 36 lines into movallow.  But this kludge
works around the bug, whereas I'm not at all sure I wouldn't break
something worse by messing with the guts of movallow.