Make fsck pay attention to clean bits once again.
I'm a bit conflicted about this.  On the one hand, I believe that fsck
should do what it's told, so if you tell it "fsck this" it should fsck
it.  But "check this filesystem" is conflated with "do boot-time
checks", and, for the latter, clean bits are a Good Thing.

Most fsck runs are probably boot-time runs, so I think practicality
wins.