Get mbuf data areas aligned to 64-bit, not 32-bit, boundaries.
This is needed for NFS booting, at a minimum, because
nfs_boot_setrecvtimo expects to be able to store a struct timeval in
one directly.

This is, arguably, broken, in that it is overlaying a struct onto a
buffer rather than doing proper serializing and deserializing.  But it
arguably isn't broken, either, given the apparently-intended uses of
mbufs, and it certainly is the least-pain path to fixing the crash.